requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.2.1
dnspython==1.15.0
enum34==1.1.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-etcd==0.4.5
pytz==2013.8
requests==2.1.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
